Fix - Order mail having Arastta hardcoded
Changed Arastta to {store_name}

TODO: Fix all those texts to fit better for each status and use. Also "Thank you for your interest in {store_name} products." should probably rather be "Thank you for your interest in {store_name}'s products." (as in "Thank you for your interest in Arastta's products."). But that is beyond this simple and quick fix to make it match ...